// src/api/models/FileListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FileInfo } from './FileInfo';
export type FileListResponse = {
    name?: string;
    path?: string;
    type?: string;
    children?: Array<FileInfo>;
};